Move documentation generation utilities to a submodule where they can be ignored for coverage purposes as they aren't used in the actual application.
