+++ help_desk_api/utils/__init__.py
@@ -1,5 +1,6 @@
-from .fields import ZendeskFields  # noqa
-from .macros import ZendeskMacros  # noqa
-from .triggers import ZendeskTriggers  # noqa
+from help_desk_api.utils.document_generation.fields import ZendeskFields  # noqa
+from help_desk_api.utils.document_generation.macros import ZendeskMacros  # noqa
+from help_desk_api.utils.document_generation.triggers import ZendeskTriggers  # noqa
+from help_desk_api.utils.document_generation.views import ZendeskViews  # noqa
+
 from .utils import get_zenpy_request_vars  # noqa
-from .views import ZendeskViews  # noqa
